use fetch

+++ server/repositories/todo.repository.ts
@@ -2,16 +2,19 @@ import { db} from "@/server/lib/db";
 import {RowDataPacket, ResultSetHeader} from "mysql2";//
 import { Todo } from "@/types/todo";
 
-export async function findAll(): Promise<Todo[]> {
-  const [rows] = await db.query<RowDataPacket[]>("SELECT * FROM todos ORDER BY id DESC");
+export async function findAll(userId: number): Promise<Todo[]> {
+  const [rows] = await db.query<RowDataPacket[]>(
+    "SELECT * FROM todos WHERE user_id = ? ORDER BY id DESC",
+    [userId]
+  );
   return rows as Todo[];
 }
 
-export async function create(title: string): Promise<Todo> {
+export async function create(title: string, userId: number ): Promise<Todo> {
   //ResultSetHeader là insertId, affectedRows, warningStatus
   const [result] = await db.execute<ResultSetHeader>(
-    "INSERT INTO todos (title, completed) VALUES (?, ?)",
-    [title, false]
+    "INSERT INTO todos (title, completed, user_id) VALUES (?, ?, ?)",
+    [title, false, userId]
   );
 
   const [rows] = await db.query<RowDataPacket[]>(
@@ -22,20 +25,20 @@ export async function create(title: string): Promise<Todo> {
    return rows[0] as Todo;
 }
 
-export async function remove(id: number): Promise<boolean> {
+export async function remove(id: number, userId: number): Promise<boolean> {
   const [result] = await db.execute<ResultSetHeader>(
-    "DELETE FROM todos WHERE id = ?",
-    [id]
+    "DELETE FROM todos WHERE id = ? AND user_id = ?",
+    [id, userId]
   );
 
   return result.affectedRows > 0;
 }
 
 
-export async function update(id: number, title: string): Promise<Todo | null> {
+export async function update(id: number, title: string, userId: number): Promise<Todo | null> {
   const [result] = await db.execute<ResultSetHeader>(
-    "UPDATE todos SET title = ? WHERE id = ?",
-    [title, id]
+    "UPDATE todos SET title = ? WHERE id = ? AND user_id = ?",
+    [title, id, userId]
   );
 
   const [rows] = await db.query<RowDataPacket[]>(
@@ -49,10 +52,10 @@ export async function update(id: number, title: string): Promise<Todo | null> {
   return rows[0] as Todo;
 }
 
-export async function toggle(id: number): Promise<Todo | null> {
+export async function toggle(id: number, userId: number): Promise<Todo | null> {
     await db.execute<ResultSetHeader>(
-        "UPDATE todos SET completed = NOT completed WHERE id = ?",
-        [id]
+        "UPDATE todos SET completed = NOT completed WHERE id = ? AND user_id = ?",
+        [id, userId]
     );
 
     const [rows] = await db.query<RowDataPacket[]>(
